Attempt a non-SIP approach
This does not work as of yet. I have modified the script so that it is
not going into a SIP location. What's failing is that essentially, the
patch.pl requires both the old and new strings to be of identical
length. So for this to work, you'd have to modify the QUARKS definition
at the top of the file to something that has the same length as the
string we are replacing.

Signed-off-by: Matthias Nott <matthias.nott@sap.com>